contex-ai-prompt-generation-backend/core_api/prompts/urls.py
from rest_framework.routers import DefaultRouter

from . import views

app_name = "prompts"
router = DefaultRouter()

router.register(r"", views.PromptsViewSet, basename="prompts")

urlpatterns = router.urls
