contex-ai-prompt-generation-backend/requirements.txt
uvicorn
openai
pydantic
django
dotenv
django-cors-headers
django-filter
django-log-request-id
django-model-utils
django-request-logging
django-storages
djangorestframework
djangorestframework-simplejwt